transx/pyproject.toml
[tool.poetry]
name = "transx"
version = "0.1.0"
description = "Video transcribing and translation tools"
authors = ["Julio Faerman <356476+faermanj@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
click = "^8.1.7"
boto3 = "^1.34.84"
dynaconf = "^3.2.5"
tenacity = "^8.2.3"
pyssml = "^0.1.3"
webvtt-py = "^0.4.6"
pyvimeo = "^1.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
transx = "transx.main:main"
